Fix rendering
